+++ R/aaa.R
@@ -1,0 +1,6 @@
+# better try
+tri <- function(expr) {
+  tryCatch(expr, error = function(e) e)
+}
+is.condition <- function(x) inherits(x, 'condition')
+is.error_cond <- function(x) is.condition(x) && inherits(x, 'error')