+++ R/routestack.R
@@ -50,8 +50,13 @@
 #'  of routes in sequence until one of the routes return `FALSE` or every route
 #'  have been passed through. `...` will be passed on to the dispatch of each
 #'  `Route` on the stack.}
-#'  \item{`on_attach(app, ...)`}{Method for use by `fiery` when attached as a
-#'  plugin. Should not be called directly.}
+#'  \item{`on_error(fun)`}{Set the error handling function. This must be a
+#'  function that accepts an `error`, `request`, and `reponse` argument. The
+#'  error handler will be called if any of the route handlers throws an error
+#'  and can be used to modify the `500` response before it is send back. By
+#'  default, the error will be signaled using `message`}
+#'  \item{`on_attach(app, on_error = NULL, ...)`}{Method for use by `fiery` when
+#'  attached as a plugin. Should not be called directly.}
 #' }
 #'
 #' @section Fiery plugin:
@@ -83,6 +88,11 @@
 #'
 #' How a `RouteStack` is attached is defined by the `attach_to` field which must
 #' be either `'request'`, `'header'`, or `'message'`.
+#'
+#' When attaching the `RouteStack` it is possible to modify how errors are
+#' handled, using the `on_error` argument, which will change the error handler
+#' set on the `RouteStack`. By default the error handler will be changed to
+#' using the `fiery` logging system if the `Fire` object supports it.
 #'
 #' @seealso [Route] for defining single routes
 #'
@@ -127,6 +137,9 @@ RouteStack <- R6Class('RouteStack',
         lapply(names(routes), function(name) {
           self$add_route(routes[[name]], name)
         })
+      }
+      private$error_fun <- function(error, request, response) {
+        message('routr error: ', conditionMessage(error))
       }
     },
     print = function(...) {
@@ -176,13 +189,26 @@ RouteStack <- R6Class('RouteStack',
         request <- as.Request(request)
       }
       for (route in private$stack) {
-        continue <- route$dispatch(request, ...)
+        continue <- tri(route$dispatch(request, ...))
+        if (is.error_cond(continue)) {
+          response <- request$respond()
+          response$status <- 500L
+          private$error_fun(continue, request, response)
+          break
+        }
         if (!continue) break
       }
       continue
     },
-    on_attach = function(app, ...) {
+    on_attach = function(app, on_error = NULL, ...) {
       assert_that(inherits(app, 'Fire'))
+      if (!is.null(app$log) && is.null(on_error)) {
+        self$on_error(function(error, request, response) {
+          app$log('error', conditionMessage(error))
+        })
+      } else if (!is.null(on_error)) {
+        self$on_error(on_error)
+      }
       if (self$attach_to == 'message') {
         assert_that(!is.null(private$path_from_message))
         app$on('message', function(server, id, binary, message, request, arg_list) {
@@ -198,6 +224,11 @@ RouteStack <- R6Class('RouteStack',
           self$dispatch(request, server = server, id = id, arg_list = arg_list)
         })
       }
+    },
+    on_error = function(fun) {
+      assert_that(is.function(fun))
+      assert_that(has_args(fun, c('error', 'request', 'response')))
+      private$error_fun <- fun
     }
   ),
   active = list(
@@ -213,6 +244,7 @@ RouteStack <- R6Class('RouteStack',
     stack = list(),
     routeNames = character(),
     attachAt = 'request',
-    path_from_message = NULL
+    path_from_message = NULL,
+    error_fun = NULL
   )
 )